Create product repo with first functionalities

+++ src/main/java/org/example/Product.java
@@ -1,4 +1,4 @@
 package org.example;
 
-public record Product(String id, String name, String category, String description, int price) {
+public record Product(String id, String name, String category, String description, double price) {
 }
